# build/generated/openapi/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Asset Library for Crypto Projects",
    "contact" : {
      "name" : "Ammer Group",
      "url" : "https://ammer.group",
      "email" : "info@ammer.group"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/assets",
    "description" : "localhost"
  } ],
  "tags" : [ {
    "name" : "Config Retrieval service",
    "description" : "Get the value for a certain config"
  } ],
  "paths" : {
    "/assets/api/admin/asset" : {
      "put" : {
        "tags" : [ "Config Retrieval service" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/admin/assets/{assetId}" : {
      "delete" : {
        "tags" : [ "Config Retrieval service" ],
        "parameters" : [ {
          "name" : "assetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/admin/network" : {
      "put" : {
        "tags" : [ "Config Retrieval service" ],
        "summary" : "Finds Pets by status",
        "description" : "Multiple status values can be provided with comma separated strings",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "string"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Successful creation of a request for signature",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
    "/assets/api/admin/networks/{networkId}" : {
      "delete" : {
        "tags" : [ "Config Retrieval service" ],
        "parameters" : [ {
          "name" : "networkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/public/changes" : {
      "get" : {
        "parameters" : [ {
          "name" : "startSequenceNumber",
          "in" : "query",
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/public/getSupportedCodecs" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/public/getSupportedPublicKeys" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/public/networks" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/public/networks/{networkId}/assets/media-assets" : {
      "get" : {
        "parameters" : [ {
          "name" : "networkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/public/networks/{networkId}/assets/native" : {
      "get" : {
        "parameters" : [ {
          "name" : "networkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/assets/api/public/networks/{networkId}/assets/smart-assets" : {
      "get" : {
        "parameters" : [ {
          "name" : "networkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  }
}